Move CryptKey namespace & injection (it isn't related to updating)
