+++ src/App/Container.php
@@ -170,6 +170,7 @@ class Container implements ContainerInterface
                         $this->container->make(DeployProcess\CompressStaticContent::class),
                         $this->container->make(DeployProcess\DisableGoogleAnalytics::class),
                         $this->container->make(DeployProcess\UnlockCronJobs::class),
+                        $this->container->make(DeployProcess\CryptKey::class),
                         /**
                          * Remove this line after implementation post-deploy hook
                          */
@@ -213,7 +214,6 @@ class Container implements ContainerInterface
                     'processes' => [
                         $this->container->make(DeployProcess\InstallUpdate\ConfigUpdate\CronConsumersRunner::class),
                         $this->container->make(DeployProcess\InstallUpdate\ConfigUpdate\DbConnection::class),
-                        $this->container->make(DeployProcess\InstallUpdate\ConfigUpdate\CryptKey::class),
                         $this->container->make(DeployProcess\InstallUpdate\ConfigUpdate\Amqp::class),
                         $this->container->make(DeployProcess\InstallUpdate\ConfigUpdate\Redis::class),
                         $this->container->make(DeployProcess\InstallUpdate\ConfigUpdate\SearchEngine::class),
